LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/naturalWidth-naturalHeight-expected.txt



PASS HTMLImageElement.prototype.naturalWidth/naturalHeight, raster image
PASS HTMLImageElement.prototype.naturalWidth/naturalHeight, raster image with width/height attributes
PASS HTMLImageElement.prototype.naturalWidth/naturalHeight, non existent image, no natural dimensions
PASS HTMLImageElement.prototype.naturalWidth/naturalHeight, non existent image with width/height attributes, no natural dimensions
FAIL HTMLImageElement.prototype.naturalWidth/naturalHeight, SVG image, no natural dimensions assert_equals: naturalWidth expected 0 but got 300
PASS HTMLImageElement.prototype.naturalWidth/naturalHeight, SVG image, width/height in pixels
FAIL HTMLImageElement.prototype.naturalWidth/naturalHeight, SVG image, width in pixels; height unspecified assert_equals: naturalWidth expected 500 but got 300
FAIL HTMLImageElement.prototype.naturalWidth/naturalHeight, SVG image, width in pixels; percentage height assert_equals: naturalWidth expected 500 but got 300
PASS HTMLImageElement.prototype.naturalWidth/naturalHeight, SVG image, width/height in pixels; viewBox
FAIL HTMLImageElement.prototype.naturalWidth/naturalHeight, SVG image, width/height unspecified; viewBox assert_equals: naturalWidth expected 0 but got 800
FAIL HTMLImageElement.prototype.naturalWidth/naturalHeight, SVG image, width in pixels; height unspecified; viewBox assert_equals: naturalWidth expected 400 but got 800
FAIL HTMLImageElement.prototype.naturalWidth/naturalHeight, SVG image, width unspecified; height in pixels; viewBox assert_equals: naturalWidth expected 400 but got 800
